Emprunt + Facturation + Notification

+++ backend/src/main/java/com/mylibrary/backend/repository/LivreRepository.java
@@ -2,4 +2,4 @@ package com.mylibrary.backend.repository;
 import com.mylibrary.backend.entitie.Livre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LivreRepository extends JpaRepository<Livre, Integer> {}
+public interface LivreRepository extends JpaRepository<Livre, Long> {}
